requirements.txt
langchain==0.3.13
langchain_community==0.3.13
langchain_core==0.3.28
langchain_groq==0.2.2
langchain_huggingface==0.1.2
python-dotenv==1.0.1
streamlit==1.41.1
bs4==0.0.1
faiss-cpu==1.9.0.post1